% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASB.R
\name{ASB}
\alias{ASB}
\title{Absolute standardized bias}
\usage{
ASB(data, confounders, treatment)
}
\arguments{
\item{data}{a dataframe or equivalent}

\item{confounders}{a vector containing the column names of the confounders to balance by}

\item{treatment}{a string indicating the column name of the treatment variable}
}
\value{
a list of objects, containing the ASB matrix and some summary statistics
}
\description{
Compute the absolute standardized bias of given confounders and return some useful statistics
}
\examples{
out = ASB(data = MARMoT_data, confounders = c("race", "age"), treatment = "hospital")

}
