% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deloof_data}
\alias{deloof_data}
\title{Data to showcase deloof and mcdeloof functions}
\format{
\subsection{\code{deloof_data}}{

A data frame with 8,450 rows and 2 columns:
\describe{
\item{comorbidities, age}{our partially ordered variables}
}
}
}
\source{
{Created in-house to serve as an example}
}
\usage{
deloof_data
}
\description{
A very simple example dataset
}
\keyword{datasets}
