% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.build_incidence}
\alias{.build_incidence}
\title{Build sparse incidence matrix for a set of pairs}
\usage{
.build_incidence(pairs, n)
}
\description{
Build sparse incidence matrix for a set of pairs
}
\keyword{internal}
