% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{.create_omics_list}
\alias{.create_omics_list}
\title{Create a list of omic matrices for JDR.}
\usage{
.create_omics_list(omics, names, sep, overlap_samples)
}
\arguments{
\item{omics}{A string vector containing filenames for omics matrices;
accepts text or .RData files; matrices must have samples as columns
and rows as features.The omics data must come from the same samples,
but some samples may have some missing values in some omics.}

\item{names}{Character vector with the names to be used for the omics;
if left NULL, the omics will just be numbered.}

\item{sep}{Character vector containing the separator used if text
files are provided.}

\item{overlap_samples}{Logical. Whether to ensure only samples with data
in all omics are kept. Some JDR methods require this.}
}
\value{
A list of omics matrices for JDR.
}
\description{
Internal function that loads omics data and
binds them into a list.
}
