% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{.filter_omics}
\alias{.filter_omics}
\title{Filter omics to only overlapping samples.}
\usage{
.filter_omics(omic_list)
}
\arguments{
\item{omic_list}{A list of omic matrices.}
}
\value{
A list of omics that are filtered to only common samples.
}
\description{
Internal function to filter omics to only overlapping samples.
This is required by some JDR methods.
}
