% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{.filter_pcs}
\alias{.filter_pcs}
\title{Filter PCA results based on a R2 threshold.}
\usage{
.filter_pcs(omic_pca, thresh, n_pcs)
}
\arguments{
\item{omic_pca}{PCA results for one omic.}

\item{thresh}{NULL or numeric between 0-1. Threshold for the R2_cum
to be used for selecting PCs; Only needed if 'PCA' is TRUE.}

\item{n_pcs}{Numeric, indicating the number of principle components to keep.
If 'thresh' is not NULL, it will ensure at least this many PCs are selected
if fewer are needed to reach the threshold.}
}
\value{
A matric of omic principle components filtered based on a
cummulative R2 threshold.
}
\description{
Internal function that filters a list of omics
based on a cummulative R2 threshold; alternatively,
it can select a specified number of principle components.
}
