% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{.merge_feat}
\alias{.merge_feat}
\alias{.merge_surv}
\title{Merge multiple columns of surv data frame}
\usage{
.merge_surv(cols, surv)
}
\arguments{
\item{cols}{Character vector containing column names.}

\item{surv}{Data frame of extracted survival features.}
}
\value{
A data frame with merged columns.
}
\description{
Merges the columns of a data frame if more than one feature
is provided for a single feature.
}
