% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.mode_val}
\alias{.mode_val}
\title{Mode of a vector (most frequent value, ties broken by first occurrence)}
\usage{
.mode_val(x)
}
\description{
Mode of a vector (most frequent value, ties broken by first occurrence)
}
\keyword{internal}
