% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{.omics_pca}
\alias{.omics_pca}
\title{Perform PCA on a list of omics.}
\usage{
.omics_pca(omic, scale)
}
\arguments{
\item{omic}{A matrix of omics data.}

\item{scale}{Logical. Whether data should be scaled prior to performing PCA.}
}
\value{
PCA results for the omic matrix.
}
\description{
Internal function that takes a list of omics and
performs PCA on them.
}
