% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_generate_pairs}
\alias{.pacmap_generate_pairs}
\title{Generate all three pair types}
\usage{
.pacmap_generate_pairs(X, n, n_neighbors, n_MN, n_FP, seed, verbose)
}
\description{
Generate all three pair types
}
\keyword{internal}
