% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_optimize}
\alias{.pacmap_optimize}
\title{Three-phase Adam optimization (vectorized via sparse incidence matrices)}
\usage{
.pacmap_optimize(
  Y,
  pair_nb,
  pair_mn,
  pair_fp,
  n,
  n_components,
  num_iters,
  lr,
  verbose
)
}
\description{
Three-phase Adam optimization (vectorized via sparse incidence matrices)
}
\keyword{internal}
