% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_preprocess}
\alias{.pacmap_preprocess}
\title{Preprocess: PCA if high-dim, else min-max normalize + center}
\usage{
.pacmap_preprocess(X, apply_pca, n_components, seed, verbose)
}
\description{
Preprocess: PCA if high-dim, else min-max normalize + center
}
\keyword{internal}
