% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_sample_FP}
\alias{.pacmap_sample_FP}
\title{Sample further pairs with rejection sampling}
\usage{
.pacmap_sample_FP(n, n_FP, nbrs, n_neighbors, seed)
}
\description{
Sample further pairs with rejection sampling
}
\keyword{internal}
