% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_sample_MN}
\alias{.pacmap_sample_MN}
\title{Sample mid-near pairs (6 random candidates, discard closest, pick 2nd closest)}
\usage{
.pacmap_sample_MN(X, n, n_MN, seed)
}
\description{
Sample mid-near pairs (6 random candidates, discard closest, pick 2nd closest)
}
\keyword{internal}
