% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.pacmap_sample_neighbors}
\alias{.pacmap_sample_neighbors}
\title{Sample nearest neighbor pairs (sort by scaled distance, keep top n_neighbors)}
\usage{
.pacmap_sample_neighbors(scaled_dist, nbrs, n, n_neighbors)
}
\description{
Sample nearest neighbor pairs (sort by scaled distance, keep top n_neighbors)
}
\keyword{internal}
