% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_build_knn}
\alias{.parc_build_knn}
\title{Build HNSW index and query k-NN}
\usage{
.parc_build_knn(data, k, n, d, distance, ef_construction, num_threads)
}
\description{
Build HNSW index and query k-NN
}
\keyword{internal}
