% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_jaccard_prune}
\alias{.parc_jaccard_prune}
\title{Compute Jaccard similarity on edges and apply global threshold}
\usage{
.parc_jaccard_prune(csr, n, jac_std_global, verbose)
}
\description{
Compute Jaccard similarity on edges and apply global threshold
}
\keyword{internal}
