% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_leiden}
\alias{.parc_leiden}
\title{Run Leiden community detection on the pruned Jaccard graph}
\usage{
.parc_leiden(
  pruned,
  n,
  jac_weighted_edges,
  partition_type,
  resolution_parameter,
  n_iter_leiden,
  seed,
  verbose
)
}
\description{
Run Leiden community detection on the pruned Jaccard graph
}
\keyword{internal}
