% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_make_csr}
\alias{.parc_make_csr}
\title{Build sparse adjacency from kNN with optional local pruning}
\usage{
.parc_make_csr(
  neighbor_array,
  distance_array,
  n,
  k,
  keep_all_local_dist,
  dist_std_local,
  verbose
)
}
\description{
Build sparse adjacency from kNN with optional local pruning
}
\keyword{internal}
