% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_merge_small}
\alias{.parc_merge_small}
\title{Merge clusters smaller than small_pop into nearest neighbor clusters}
\usage{
.parc_merge_small(labels, neighbor_array, small_pop, time_smallpop, verbose)
}
\description{
Merge clusters smaller than small_pop into nearest neighbor clusters
}
\keyword{internal}
