% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc.R
\name{.parc_split_toobig}
\alias{.parc_split_toobig}
\title{Split clusters that are too big by recursive sub-clustering}
\usage{
.parc_split_toobig(
  labels,
  data,
  neighbor_array,
  distance_array,
  n,
  k,
  too_big_factor,
  keep_all_local_dist,
  dist_std_local,
  jac_std_global,
  jac_weighted_edges,
  partition_type,
  resolution_parameter,
  n_iter_leiden,
  seed,
  small_pop,
  time_smallpop,
  distance,
  hnsw_param_ef_construction,
  num_threads,
  verbose
)
}
\description{
Split clusters that are too big by recursive sub-clustering
}
\keyword{internal}
