% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacmap.R
\name{.sample_reject}
\alias{.sample_reject}
\title{Rejection sampler: pick n_samples from 1:maximum, excluding 'exclude'}
\usage{
.sample_reject(n_samples, maximum, exclude, max_tries = 1000L)
}
\description{
Rejection sampler: pick n_samples from 1:maximum, excluding 'exclude'
}
\keyword{internal}
