% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{getMetadata}
\usage{
getMetadata(localfilepath = ".")
}
\value{
A message indicating the location of the metadata file.
}
\description{
This function copies the metadata file from the package to a user-defined location.
}
\examples{
\dontrun{
marmot(metadata = "~/Desktop/Flow_Data/MARMOT_metadata.xlsx", name = "Study Name", render = FALSE)
}
}
\author{
Peter Leary
}
