% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_interface.R
\name{install_p4r_env}
\alias{install_p4r_env}
\title{Install Python Environment for MARMOT}
\usage{
install_p4r_env()
}
\value{
A message indicating success.
}
\description{
Installs a conda environment called \code{p4r} using the \code{environment.yml}
specification bundled with the package.
}
\examples{
\dontrun{
install_p4r_env()
}
}
