% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_io.R
\name{load_h5ad_to_env}
\alias{load_h5ad_to_env}
\title{Load h5ad Data into Environment}
\usage{
load_h5ad_to_env(h5ad_path, envir = .GlobalEnv)
}
\arguments{
\item{h5ad_path}{Path to the marmot_results.h5ad file}

\item{envir}{Environment to load objects into (default: .GlobalEnv)}
}
\description{
Reads a MARMOT h5ad file and reconstructs all pipeline R objects
in the specified environment.
}
