% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_to_feat_weights.r
\name{map_weights_to_features}
\alias{map_weights_to_features}
\title{Map weights to features.}
\usage{
map_weights_to_features(
  jdr_methods = c("MOFA", "JIVE", "RGCCA", "MCIA"),
  fct_list,
  pca_weights,
  omics = NULL
)
}
\arguments{
\item{jdr_methods}{Character vector specifying one more more JDR methods
to be used. Should be from \code{c("MOFA", "JIVE", "RGCCA", "MCIA")}.}

\item{fct_list}{A list of factorissations. Expects output of
\code{\link{run_jdr}}, or one of the individual functions.}

\item{pca_weights}{A pcaMethods output corresponding to the data
imputted for JDR. See \code{\link{prepare_data}}.}

\item{omics}{Names of the omics for which this should be done.
If NULL, it will be done on all omics}
}
\description{
This function maps factor weights of princiople components
back to individual features (e.g. genes). Performing PCA on omics data prior
to JDR can have several advantages. However, one of the strength of JDR
approaches is being able to identify which individual features contribute
most to the factorisation. In order achieve this when performing PCA first,
we must multiply the matrix of PCA loadings to the matrix of factor weights.
This only works with JDR methods that are linear. For details, check \link{cite my paper} # nolint
}
