% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MARMoT_data}
\alias{MARMoT_data}
\title{Data to showcase MARMoT and ASB functions}
\format{
\subsection{\code{MARMoT_data}}{

A data frame with 8450 rows and 3 columns:
\describe{
\item{hospital}{our treatment}
\item{race, age}{our confounders}
}
}
}
\source{
{Created in-house to serve as an example}
}
\usage{
MARMoT_data
}
\description{
An example dataset where our treatment is the hospital structure.
}
\keyword{datasets}
