% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdeloof.R
\name{mcdeloof}
\alias{mcdeloof}
\title{Average rank with Deloof's approximation (parallel computing)}
\usage{
mcdeloof(comparable_data, n.cores)
}
\arguments{
\item{comparable_data}{a dataframe or equivalent, which columns are ordered factors or numeric variables}

\item{n.cores}{number of cores to be used}
}
\value{
a vector containing the average rank of all the observations (it's recommended to normalized it before use)
}
\description{
Compute the average with using Deloof's approximation, using parallel computing; it can be used only on Linux and Mac systems!
}
\examples{
AR = mcdeloof(deloof_data, n.cores = 1)

}
