% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\docType{data}
\name{p4r_env}
\alias{p4r_env}
\title{Basilisk Python environment for PARC and PaCMAP}
\format{
A \code{BasiliskEnvironment} object.
}
\usage{
p4r_env
}
\description{
A \code{\link[basilisk]{BasiliskEnvironment}} providing the \code{p4r} conda
environment used by MARMOT for PARC clustering and PaCMAP dimensionality
reduction. The environment is created automatically on first use via
\pkg{basilisk} and managed in the user's package cache.
}
\keyword{datasets}
