% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.r
\name{perform_gsea}
\alias{perform_gsea}
\title{Run GSEA based on differential analysis results}
\usage{
perform_gsea(
  diff_results,
  limma = TRUE,
  gene_set,
  save_file = TRUE,
  file_name = NULL,
  ...
)
}
\arguments{
\item{diff_results}{A data frame containing either a limma toptable or the
output of \code{link{differential_analysis}}.}

\item{gene_set}{A .gmt file containing gene sets of interest or a list of
gene sets to test.}

\item{...}{Any other parameters accepted by fgsea.}
}
\value{
Results of GSEA
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) on
results of differential analysis based on JDR factors. See
\code{\link{differential_analysis}} for details.
}
\seealso{
\code{\link{fgsea::fgsea}}
}
