% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotUMAPNice}
\alias{plotUMAPNice}
\title{plotUMAPNice}
\usage{
plotUMAPNice(
  df = umapDF,
  colour_by = "condition",
  show_density = FALSE,
  size = 0.7,
  alpha = 1,
  text_size = 12,
  facet_by = NULL,
  show_grids = FALSE,
  name = "Normalised\\nExpression\\n",
  show_cluster_labels = FALSE,
  cluster_labels = "cluster_id",
  dimRedMethodToUse = dimRedMethodToUse,
  useMarkers = useMarkers,
  coloursList = coloursList,
  viridisColour = viridisColour
)
}
\value{
A ggplot object
}
\description{
This is our little custom function for drawing nice and customisable DR plots
}
\author{
Peter Leary
}
