% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for JDR.}
\usage{
prepare_data(
  omics,
  names = NULL,
  overlap_samples = TRUE,
  pca = TRUE,
  thresh = 0.85,
  n_pcs = 20,
  save_pca = TRUE,
  file_name = NULL,
  scale = TRUE
)
}
\arguments{
\item{omics}{A string vector containing filenames for omics matrices;
accepts .RData files or any format accepted by fread; matrices must have
samples as columns and rows as features. The omics data must come from
the same samples, but some samples may have some missing values in some
omics.}

\item{names}{Character vector with the names to be used for the omics;
if left NULL, the omics will just be numbered.}

\item{overlap_samples}{Logical. Whether to ensure only samples with data
in all omics are kept. Some JDR methods require this.}

\item{pca}{Logical. Whether PCA should be performed on the omics.}

\item{thresh}{NULL or numeric between 0-1. Threshold for the R2_cum
to be used for selecting PCs; Only needed if 'PCA' is TRUE.}

\item{n_pcs}{Numeric, indicating the number of principle components to keep.
If 'thresh' is not NULL, it will ensure at least this many PCs are selected
if fewer are needed to reach the threshold.}

\item{save_pca}{Logical. Whether to save a copy of the unfiltered PCA results
as an RData file.}

\item{file_name}{Character string specifying the file name for the
unfiltered PCA results. If left NULL, a generic name will be assigned.}

\item{scale}{Logical. Whether data should be scaled prior to performing PCA.}
}
\value{
A list of omics ready for JDR.
}
\description{
This function prepares omics data for joint dimensionality
reduction.
}
