% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_interface.R
\name{python_interface}
\alias{python_interface}
\title{Initialise Python Environment for MARMOT}
\description{
Helper functions to manage and use the Python environment required by MARMOT.
This includes automatic sourcing of the bundled \code{f_pacmap.py} and \code{f_parc.py} scripts.
}
\author{
Peter Leary
}
