% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_factorisations.r
\name{run_jdr}
\alias{run_jdr}
\title{Run joint dimensionality reduction on multi-omics data.}
\usage{
run_jdr(
  omic_list,
  samples_overlap = TRUE,
  pca = TRUE,
  jdr_methods = c("MOFA", "JIVE", "RGCCA", "MCIA"),
  n_fct = 5,
  seed = 42,
  convergence = "slow",
  use_basilisk = TRUE
)
}
\arguments{
\item{omic_list}{A list of omic matrices.
Should be output of \code{link{prepare_data}}.}

\item{samples_overlap}{Logical. Whether a sample overlap was performed
between the omics during data preparation. See \code{\link{prepare_data}}.
If FALSE, omics will be filtered to only common samples for JDR methods that
require it. This may result in different sample sets being used with
different methods.}

\item{pca}{Logiacal. Whether PCA was performed on the data.
See \code{\link{prepare_data}}.}

\item{jdr_methods}{Character vector specifying one more more JDR methods
to be used. Should be from \code{c("MOFA", "JIVE", "RGCCA", "MCIA")}.}

\item{n_fct}{Integer. Number of factors for the factorisations.}

\item{seed}{Seed for factorisation.}

\item{convergence}{Only needed if MOFA2 is used. Determines the convergence
mode. Can be one of c("slow", "medium", "fast").}

\item{use_basilisk}{Logical. Whether basilisk should be used to automatically
set up conda environment. Only needed when MOFA2 is used.}

\item{...}{Any other parameters that can be passed to any used functions
from the JDR packages. See respective package documentation for further details. #nolint}
}
\value{
A list of factorisations. Each element is the factorisation
based on one JDR method.
}
\description{
Run joint dimensionality reduction on multi-omics data.
}
