% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_factorisations.r
\name{run_jive}
\alias{run_jive}
\title{Run JDR with JIVE.}
\usage{
run_jive(omic_list, n_fct)
}
\arguments{
\item{omic_list}{A list of omic matrices.
Should be output of \code{link{prepare_data}}.}

\item{n_fct}{Integer. Number of factors for the factorisations.}
}
\value{
A trained JIVE model.
}
\description{
Performs JRD on a list of omic matrices with JIVE.
}
