% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_factorisations.r
\name{run_mofa2}
\alias{run_mofa2}
\title{Run JDR with MOFA2.}
\usage{
run_mofa2(omic_list, n_fct, seed, convergence, use_basilisk, ...)
}
\arguments{
\item{...}{Any other parameters that can be passed to MOFA2 functions.
See documentation of the MOFA2 package for details.}
}
\value{
A trained MOFA2 model.
}
\description{
Performs JDR on a list of omic matrices with MOFA2.
}
