% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_interface.R
\name{source_p4r_scripts}
\alias{source_p4r_scripts}
\title{Source bundled Python scripts}
\usage{
source_p4r_scripts()
}
\value{
Nothing, but defines Python functions in the R environment.
}
\description{
Loads \code{f_pacmap.py} and \code{f_parc.py} into the current R session via \pkg{reticulate}.
}
\examples{
\dontrun{
source_p4r_scripts()
}
}
