% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_analysis.r
\name{surv_association}
\alias{surv_association}
\title{Calculate factor association with survival.}
\usage{
surv_association(factors, surv, univariate = TRUE)
}
\arguments{
\item{factors}{Data frame containing factors to be associated with survival.
Samples should be rows and factors columns.}

\item{surv}{A data frame with survival information. Expects output from
\code{\link{prepare_surv}}.}

\item{univariate}{Logical, whether to run a univariate coxph for each
factor (TRUE) or a multivariate coxph for all together (FALSE).
Default is TRUE.}
}
\value{
A list with either one multivariate coxph model or multiple
univariate coxph models.
}
\description{
This function takes as input computed factorisations from JDR
models (see \code{\link{run_jdr}}) and calculates the association to survival
of the factors. There are options for either univariate cox regression for
each factor, or multivariate cox regression for multiple factors.
}
