% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetadataToFCSFolder.R
\name{addMetadataToFCSFolder}
\alias{addMetadataToFCSFolder}
\title{addMetadataToFCSFolder}
\usage{
addMetadataToFCSFolder(FCS_folder = ".", name = NULL, overwrite = FALSE)
}
\arguments{
\item{FCS_folder}{Path to the folder containing FCS files (default ".")}

\item{name}{Name suffix for the metadata file (default NULL)}

\item{overwrite}{Logical; overwrite existing file? (default FALSE)}
}
\value{
A message indicating the location of the metadata file.
}
\description{
This function copies the metadata file from the package to a user-defined location. Ideally the folder where the gated FCS files live.
}
\examples{
\dontrun{
addMetadataToFCSFolder("Files/MARMOT_Metadata.xlsx")
}
}
\author{
Peter Leary
}
