% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deloof.R
\name{deloof}
\alias{deloof}
\title{Average rank with Deloof's approximation}
\usage{
deloof(comparable_data)
}
\arguments{
\item{comparable_data}{A dataframe or equivalent, which columns are ordered factors or numeric
variables.}
}
\value{
A vector containing the average rank of all observations (it is recommended
to normalized it before use).
}
\description{
Compute the average with using Deloof's approximation.
}
\examples{
AR = deloof(deloof_data)

}
\references{
Caperna, G., 2019. Approximation of AverageRank by means of a formula.
https://doi.org/10.5281/zenodo.2565699
}
