% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.r
\name{differenitial_analysis}
\alias{differenitial_analysis}
\alias{differential_analysis}
\title{Differential analysis based on JDR factors.}
\usage{
differential_analysis(
  omic,
  factor,
  clin,
  feat,
  covariates = NULL,
  limma = TRUE,
  survival = TRUE
)
}
\arguments{
\item{omic}{A matrix of omics data. Columns should be samples and features
should be rows.}

\item{factor}{A dataframe with the factor based on which to determine groups.}

\item{clin}{A dataframe of clinical data.}

\item{feat}{Character string with the clinical feature with which the factor
of interest is associated. Must be a column name in \code{clin}.}

\item{covariates}{Optional. A character vector with names of covariates for
which to correct. Only applies if limma is used.}

\item{limma}{Whether limma should be used for the differential analysis.
If FALSE, a wilcoxon signed rank test will be used instead. Default is TRUE.}

\item{survival}{Logical. Whether the feature of interest is survival. This
will determine how the groups are divided.}
}
\description{
This function performs differential analysis on omics data. It
defines the groups based on given JDR factors associated with a given
clinical feature.
}
