% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.r
\name{differenitial_analysis}
\alias{differenitial_analysis}
\alias{differential_analysis}
\title{Differential analysis based on JDR factors.}
\usage{
differential_analysis(
  omic,
  factor,
  surv,
  clin = NULL,
  covariates = NULL,
  limma = TRUE,
  minprop = 0.1
)
}
\arguments{
\item{omic}{A matrix of omics data. Columns should be samples and features
should be rows.}

\item{factor}{A dataframe with the factor based on which to determine groups.}

\item{surv}{A data frame with survival information. Expects output from
\code{\link{prepare_surv}}.}

\item{clin}{Optional. A dataframe with clinical features for which to correct
if using limma.}

\item{covariates}{Optional. A character vector with names of covariates for
which to correct. Only applies if limma is used.}

\item{limma}{Whether limma should be used for the differential analysis.
If FALSE, a wilcoxon signed rank test will be used instead. Default is TRUE.}

\item{minprop}{Numeric between c(0,1), indicating the minimum proportion of
samples per group.}
}
\description{
This function performs differential analysis on omics data. It
defines the groups based on given JDR factors associated with a given
clinical feature.
}
