% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Extract the template MARMOT metadata file to a specific folder for easy editing in Excel.}
\usage{
getMetadata(folder = NULL)
}
\description{
Extract the template MARMOT metadata file to a specific folder for easy editing in Excel.
}
\examples{
\dontrun{
getMetadata(folder = "~/Desktop/MARMOT/")
}
}
