% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Extract the metadata file to a specific folder}
\usage{
getMetadata(folder = NULL)
}
\description{
Extract the metadata file to a specific folder
}
\examples{
\dontrun{
getMetadata(folder = "~/Desktop/MARMOT/")
}
}
