% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARMoT.R
\name{MARMoT}
\alias{MARMoT}
\title{MARMoT balancing method}
\usage{
MARMoT(data, confounders, treatment, reference = "median", n.cores = 1)
}
\arguments{
\item{data}{a dataframe or equivalent}

\item{confounders}{a vector containing the column names of the confounders to balance by}

\item{treatment}{a string indicating the column name of the treatment variable}

\item{reference}{"median" default, the statistic used to determine the reference frequencies in the balancing process}

\item{n.cores}{1 default, number of cores to be used (Linux and Mac systems only!); if a number grater than 1 is specified the function will use a parallelized version of the deloof approximation}
}
\value{
a list of objects, also containing the balanced dataset with the same structure of the input dataset
}
\description{
Matching on poset-based average rank for multiple treatments (MARMoT)
}
\examples{
out = MARMoT(data = MARMoT_data, confounders = c("race", "age"),
      treatment = "hospital", n.cores = 1)

}
