% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot.R
\name{marmot}
\alias{marmot}
\title{The main MARMOT pipeline. Use to generate a customised MARMOT pipeline script based on the edited metadata file, and render if required.}
\usage{
marmot(metadata = NULL, name = "Title", render = FALSE)
}
\description{
The main MARMOT pipeline. Use to generate a customised MARMOT pipeline script based on the edited metadata file, and render if required.
}
\examples{
\dontrun{
marmot(metadata = "FC_metadata.xlsx", name = "Study Name", render = FALSE)
}
}
