% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot.R
\name{marmot}
\alias{marmot}
\title{marmot}
\usage{
marmot(metadata = NULL, name = "Title", render = FALSE)
}
\arguments{
\item{metadata}{Path to the MARMOT metadata Excel file (.xlsx)}

\item{name}{Name for this analysis run (used in output filenames)}

\item{render}{Logical; if TRUE, render the pipeline report (default FALSE)}
}
\value{
A results folder in the directory of the metadata containing an HTML report, and a folder with all resulting PDFs, Excel files, and R data files.
}
\description{
The main MARMOT pipeline. Use to generate a customised MARMOT pipeline script based on the edited metadata file, and render if required.
}
\examples{
\dontrun{
marmot(metadata = "~/Desktop/Flow_Data/MARMOT_metadata.xlsx", name = "Study Name", render = FALSE)
}
}
\author{
Peter Leary
}
