% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot.R
\name{marmot}
\alias{marmot}
\title{start main marmot pipeline}
\usage{
marmot(metadata = NULL, name = "MARMOT Flow Cytometry Pipeline v1.0.0")
}
\description{
start main marmot pipeline
}
\examples{
\dontrun{
marmot(metadatafile)
}
}
