% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmot.R
\name{marmot}
\alias{marmot}
\title{start main marmot pipeline}
\usage{
marmot(metadata = NULL, name = "Title", render = FALSE)
}
\description{
start main marmot pipeline
}
\examples{
\dontrun{
marmot(metadata = "FC_metadata.xlsx", name = "Study Name", render = FALSE)
}
}
