% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdeloof.R
\name{mcdeloof}
\alias{mcdeloof}
\title{Average rank with Deloof's approximation (parallel computing)}
\usage{
mcdeloof(comparable_data, n.cores)
}
\arguments{
\item{comparable_data}{A dataframe or equivalent, which columns are ordered factors or numeric
variables.}

\item{n.cores}{Number of cores to be used.}
}
\value{
A vector containing the average rank of all the observations (it is
recommended to normalized it before use).
}
\description{
Compute the average with using Deloof's approximation, using parallel
computing; it can be used only on Linux and Mac systems!
}
\examples{
AR = mcdeloof(deloof_data, n.cores = 1)

}
\references{
Caperna, G., 2019. Approximation of AverageRank by means of a formula.
https://doi.org/10.5281/zenodo.2565699
}
