% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_surv}
\alias{prepare_surv}
\title{Extract survival information from clinical data.}
\usage{
prepare_surv(
  clinical,
  feature_names,
  vital_status_values = c(alive = FALSE, dead = TRUE),
  keep_nas = FALSE
)
}
\arguments{
\item{clinical}{Character string containing path to a file with clinical
information.The expected format of clinical data is with rows as samples
and columns as features.}

\item{feature_names}{Named list containing the feature names for
the survival features of interest. Two feature names can be
provided per element. If so, the first feature will be used unless a sample
has NA values for that feature, in which case it will check for values in the
second feature. See examples for details. The list must be named as follows:
list(sample_id = "sample_id", vital_status = "vital_status",
time_to_event = "time_to_event")}

\item{vital_status_values}{A named vector of length two indicating how vital
status is recorded in the clinical data. This will be converted to logical,
where alive = FALSE and dead = TRUE. Default is c(alive = FALSE, dead = TRUE)
NAs will be preserved.}

\item{keep_nas}{Logical. If samples with NA values in any survival feature
should be kept. Not recommended for most downstream analysis.}
}
\value{
A data frame containing survival information.
}
\description{
Function that extracts survival information from a larger
dataframe of clinical features and formats it for downstream analysis.
}
