% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_factorisations.r
\name{run_jive}
\alias{run_jive}
\title{Run JDR with JIVE.}
\usage{
run_jive(omic_list, n_fct, ...)
}
\value{
A trained JIVE model.
}
\description{
Performs JRD on a list of omic matrices with JIVE.
}
