% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_factorisations.r
\name{run_rgcca}
\alias{run_rgcca}
\title{Run JDR with RGCCA.}
\usage{
run_rgcca(omic_list, n_fct)
}
\arguments{
\item{omic_list}{A list of omic matrices.
Should be output of \code{link{prepare_data}}.}

\item{n_fct}{Integer. Number of factors for the factorisations.}
}
\value{
A trained RGCCA model.
}
\description{
Performs JDR on a list of omic matrices using RGCCA.
}
