% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_io.R
\name{save_parquet_data}
\alias{save_parquet_data}
\title{Save All Pipeline Data to Parquet}
\usage{
save_parquet_data(qs_dir, envir = .GlobalEnv)
}
\arguments{
\item{qs_dir}{Path to the R_files output directory}

\item{envir}{Environment containing the pipeline variables (default: .GlobalEnv)}
}
\description{
Decomposes all pipeline R objects into tabular Parquet files within a
`parquet/` subdirectory of the R_files output directory.
}
