% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMarmot.R
\name{shinyMarmot}
\alias{shinyMarmot}
\title{shinyMarmot}
\usage{
shinyMarmot(marmot_output = NA, demo = NA)
}
\arguments{
\item{marmot_output}{Path to the pipeline output directory (R_files folder)}

\item{demo}{Logical; if TRUE, launch with bundled demo data (default NA)}
}
\value{
An interactive shiny app session in console.
}
\description{
Start the shinyMarmot application locally using results generated from the main MARMOT pipeline.
}
\examples{
\dontrun{
shniyMarmot(marmot_output = "~/Desktop/Flow_Data/Results_2025-03-10_11_01_01/R_files")
}
}
\author{
Peter Leary
}
