% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMarmot.R
\name{shinyMarmot}
\alias{shinyMarmot}
\title{Start the shinyMarmot application locally using results generated from the main MARMOT pipeline.}
\usage{
shinyMarmot(marmot_output = NA, demo = NA)
}
\description{
Start the shinyMarmot application locally using results generated from the main MARMOT pipeline.
}
\examples{
\dontrun{
marmotOut <- system.file("examples/R_files/", package = "shinyMarmot")
shniyMarmot(marmot_output)
}
}
