% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMarmot.R
\name{shinyMarmot}
\alias{shinyMarmot}
\title{start shinyMarmot application}
\usage{
shinyMarmot(marmot_output = NA, demo = NA)
}
\description{
start shinyMarmot application
}
\examples{
\dontrun{
marmotOut <- system.file("examples/R_files/", package = "shinyMarmot")
shniyMarmot(marmot_output)
}
}
